Add first iteration of token sets that replicate tailwind-css naming styles and values for our design system.
See "tailwind-primitives" token set for the constants as defined in tailwind 4.3. On top of this we built a layer with which the tokens can be affined transformed.  Mainly to fulfill the purpose of being able to quickly, globally change spacing, etc. .See "Foundation" and "Theme config: Base" token sets for this.

Co-authored-by: athorne-dev athorne-dev@users.noreply.github.com